+++ Server/src/mx_registration.c
@@ -1,0 +1,5 @@
+#include "Chat.h"
+
+void mx_registration(int sock) {
+
+}
